+++ .env
@@ -1,9 +1,9 @@
 DB_USER=test_user
 DB_PASSWORD=test1234
 DB_HOST=db-server
-DB_PORT=3306
+DB_PORT=3308
 DB_DATABASE=laundrydo
-DB_ROOT_PASSWORD=test1234
+# DB_ROOT_PASSWORD=test1234
