+++ .gitignore
@@ -4,3 +4,5 @@ __pycache__
 **/*.DS_Store
 mysql/
 db/
+*.db
+.env
