+++ src/infrastructure/fastapi/app.py
@@ -5,7 +5,6 @@ from typing import List, Dict, Annotated
 from datetime import datetime
 
 from fastapi import FastAPI, Query, Body, Depends
-from pydantic import BaseModel
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session, Session
@@ -24,9 +23,13 @@ from apscheduler.schedulers.background import BackgroundScheduler
 
 from src.application.unit_of_work import SqlAlchemyUnitOfWork
 
+from src.infrastructure.fastapi.routes import user_router, order_router
+
+
 app = FastAPI()
 
-
+app.include_router(user_router.router)
+app.include_router(order_router.router)
 
 # dependency
 MEMORY_SESSION = 'sqlite:///:memory:'
@@ -97,9 +100,6 @@ def get_db() :
         print('is db down?')
 
 
-# def print_hi() :
-#     print('hi')
-
 @app.on_event('startup')
 def init_monitor():#session : Session = Depends(get_db)) :
     ## listening on db
@@ -119,68 +119,9 @@ def init_monitor():#session : Session = Depends(get_db)) :
 def shutdown() :
     pass
 
-
 @app.get('/')
 async def root() :
     return {'LaundryDo' : 'Welcome'}
-
-@app.get('/users/{userid}', response_model = schemas.Order)
-async def request_orderlist(userid : str, session : Session = Depends(get_db)) :
-    uow = SqlAlchemyUnitOfWork(session)
-    with uow :
-        return uow.orders.get_by_userid(userid=userid)
-
-
-@app.get('/users/{userid}/orders/{orderid}')
-async def request_order_info(userid : str, orderid : str, session : Session = Depends(get_db)) -> schemas.Order : ## TODO : orders only be accessible for one user.
-    '''
-    request estimate time for order in process. if order is done or cancelled, return 0.
-    '''
-    uow = SqlAlchemyUnitOfWork(session)
-    with uow :
-        return uow.orders.get_by_orderid(orderid = orderid)
-
-
-@app.post('/users/{userid}/orders', response_model = schemas.Order)
-async def request_order(userid : str, order : Annotated[ schemas.Order, 
-            Body(
-                examples = [
-                    {   
-                        'userid' : '[userid]',
-                        "orderid" : '[userid]-order-[num]',
-                        "description" : "세탁 요청한 옷들의 리스트가 담긴 주문 정보",
-                        'clothes_list' : [{
-                            "clothesid" : "흰티셔츠",
-                            "label" : "드라이클리닝",
-                            "volume" : 3,
-                                }
-                            ],
-                        "received_at" : datetime(2023, 7, 21, 10, 11),
-                    }
-                ]
-            )
-        ], 
-        session : Session = Depends(get_db)
-    ) :
-    print(order.json())
-    uow = SqlAlchemyUnitOfWork(session)
-    with uow :
-        services.request_order(uow,**dict(order))
-        uow.commit()
-
-    return order
-
-
-@app.put('/users/{userid}/orders/{orderid}')
-async def cancel_order(userid : str, orderid : str, session : Session = Depends(get_db)) :#-> schemas.Order :
-    uow = SqlAlchemyUnitOfWork(session)
-    
-    with uow :
-        order = services.cancel_order(uow, userid, orderid)
-        order = schemas.Order.model_validate(order)
-        order.commit()
-
-    return order
